+++ CHANGELOG.md
@@ -72,6 +72,7 @@ FEATURES
   - You can now use a Ledger with `gaiacli --ledger` for all key-related commands
   - Ledger keys can be named and tracked locally in the key DB
 * [gaiacli] added an --async flag to the cli to deliver transactions without waiting for a tendermint response
+* [gaiacli] improve error messages on `send` and `account` commands
 
 IMPROVEMENTS
 * bank module uses go-wire codec instead of 'encoding/json'
@@ -100,6 +101,7 @@ BUG FIXES
 * \#1353 - CLI: Show pool shares fractions in human-readable format
 * \#1258 - printing big.rat's can no longer overflow int64
 * \#887  - limit the size of rationals that can be passed in from user input
+* \#1461 - CLI tests now no longer reset your local environment data
 
 ## 0.19.0